+++ instascrape/scrapers/__init__.py
@@ -10,3 +10,4 @@ from instascrape.scrapers.comment import *
 from instascrape.scrapers.location import *
 from instascrape.scrapers.reel import *
 from instascrape.scrapers.igtv import *
+from instascrape.scrapers.scrape_tools import *
